ref(backup): Split dependency function into own file
We will re-use this functionality in multiple places, so its best to
move it now.

Issue: getsentry/team-ospo#171
